+++ apsw/unicode.py
@@ -581,7 +581,8 @@ def text_wrap(
     :param tabsize: Tab stop spacing as tabs are expanded
     :param hyphen: Used to show a segment was broken because it was wider than ``width``
     :param combine_space: Leading space on each (indent) is always preserved.  Other spaces where
-          multiple occur are combined into one space.
+          multiple occur are combined into one space.  If tabs matter then you'll want this
+          off.
     :param invalid: If invalid codepoints are encountered such as control characters and surrogates
           then they are replaced with this.
 
@@ -592,9 +593,16 @@ def text_wrap(
     :func:`apsw.ext.format_query_table` uses this method to ensure
     each column is the desired width.
     """
+    text = expand_tabs(text, tabsize, invalid)
+
     hyphen_width = text_width(hyphen)
 
-    text = expand_tabs(text, tabsize, invalid)
+    if hyphen_width:
+        # we don't need a hyphen if the text already fits in one line
+        # which avoids truncating indent unnecessarily
+        if text_width(text) <= width:
+            hyphen = ""
+            hyphen_width = 0
 
     for line in split_lines(text):
         accumulated: list[str] = []
@@ -605,8 +613,12 @@ def text_wrap(
             if indent is None:
                 indent = " " * (len(segment) - len(segment.lstrip(" "))) if segment[0] == " " else ""
                 if len(indent) >= width - hyphen_width:
-                    # make space for double width char if indent wider than width
-                    indent = indent[: max(0, width - hyphen_width - 2)]
+                    # make space for widest char if indent wider than
+                    # width but only if we have to hyphenate. issue 600
+                    # and its dedicated tests exercise this
+                    widths = [text_width(gc) for gc in grapheme_iter(line[len(segment) :])]
+                    widest = max(widths) if widths else 0
+                    indent = indent[: max(0, width - (hyphen_width if len(widths) > 1 else 0) - widest)]
                 accumulated = [indent]
                 line_width = len(indent)
                 if line_width:
@@ -637,7 +649,7 @@ def text_wrap(
                     if desired < 1:
                         hyphen_out = ""
                         desired = width - line_width
-                    seg_width, substr = text_width_substr(segment, desired)
+                    seg_width, substr = text_width_substr(segment, desired) if desired else (0, "")
                     if seg_width == 0:
                         # the first grapheme cluster is wider than desired so
                         # we will display '*' instead for that first grapheme cluster
@@ -663,10 +675,9 @@ def text_wrap(
             if segment:
                 accumulated.append(segment)
             line_width += seg_width
-        if len(accumulated) == 1:
-            # only indent
-            yield " " * width
-        else:
+        if len(accumulated) != 1:
+            # length 1 means it is only indent which we don't output
+            # as last line
             yield "".join(accumulated) + " " * (width - line_width)